Enemies, Rounds, and Everything, Oh my!
This commit contains a load of features, which are summarized below.

Enemies:
These are things which damage the player when they collide. As of now, all enemies are coded as having a circular hitbox. There is only one type of enemy so far: the bullet. This is a basic projectile which moves in a straight line.

Rounds:
These are a series of enemies released over time. There is currently one type of round, the Straight. This round fires a projectile every y seconds from a given position in a given direction and speed.

Misc:
The player is currently moved using the mouse. There is, so far, no time slowing mechanic.