Dynamic NeuroSky use and Attention
Use of the Neurosky is now modular - if the headset is connected, it will be used. If not, it will set the attention to a central value.

Attention is displayed in the bottom right, and now modulates the time multiplier. High attention corresponds to lower time multiplier, and vis versa.