Updated Attention, Music, and Bugfix
Attention is now measured differently. Attention now uses the calibration phase data, with which it sets a threshold value. This threshold is a miniumum attention value, past which, time is slowed linearly.

Music now is not affected by the time slowing, because it was a bad listening experience.

The erroneous requirement in the requirements.txt file is now removed.